Fix Gradle brokenness on any system that does happen to be my own.
Everyone who attempts to build this project besides me encounters the
following unfortunate error messages:

    Error: Could not find or load main class org.gradle.wrapper.GradleWrapperMain
    Caused by: java.lang.ClassNotFoundException: org.gradle.wrapper.GradleWrapperMain

As it turns out, the reason for this is...gradle-wrapper.jar.  Specifically,
it needs to be checked into version control, but the gitignore created by the
libGDX project generation wizard ignored the entire ./gradle directory without
carving out an exception.

So we do it by hand.

Thanks to
https://support.snyk.io/hc/en-us/articles/360007745957-Snyk-test-Could-not-find-or-load-main-class-org-gradle-wrapper-GradleWrapperMain
for the suggestion.