Re-ran the game setup, but with a different wizard: GDX-LiftOff.
Our specific motivation for this was making sure that we could build the
project *with* the Box2dlights third-party lighting library, as lighting is
going to eventually factor into the game design.  The original gdx-setup.jar
wizard created broken HTML builds when box2dlights was part of the project, so
we found this instead:

https://github.com/tommyettinger/gdx-liftoff/

This is an alternative game setup wizard that is more up-to-date (it was last
released a couple of weeks ago), supports more third-parties, and was
specifically designed to get right what gdx-setup.jar got wrong.

It wasn't perfect, though: see the next commit.