I think I've gotten player rendering to work.
The main challenge with this commit is staging just the minimal set of changes
needed to make that happen.  I'll try to document what we've learned during
today's meeting.

The key seems to have been the call to camera.project() in order to convert
world space coordinates to screen space coordinates.  Right now, these are
always in the center of the viewport (which makes sense, seeing as how the
camera is always following the player.)  Restricting the camera when it
approaches world boundaries will allow the player to move near world corners
without "scrolling"; that will come next.